selinux: Allocate mem for contexts on context_struct_to_string() calls
It seems when arter was optimizing this method in 564ccf6 he just forgot
to update other existing calls to it with the new 4th required "alloc"
boolean argument.

I noticed this in 501b7b7 when I started testing compiling the kernel
with AppArmor and SELinux at the same time and 13 new errors popped up.